code/requirements.txt
wget~=3.2
gitpython~=3.1.27
opencv-python~=4.6.0.66
Send2Trash~=1.8.0
futures3~=1.0.0
ffmpeg-python~=0.2.0